+++ alexflipnote/errors.py
@@ -1,5 +1,6 @@
 # source: https://github.com/BlistBotList/blist-wrapper/blob/bc0c0fe9afbea39993ccfa8b6d633c2b5be634c8/blist/errors.py
 
+
 class AlexFlipnoteException(Exception):
     pass